agenttel-instrument/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agenttel-instrument"
version = "0.1.0"
description = "MCP server for AI-driven instrumentation of applications"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "agenttel-cli>=0.1.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "aiohttp>=3.9.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp-proto-http>=1.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
agenttel-instrument = "agenttel_instrument.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
